API for list all suppliers
I added a line of code in the app.js file: const suppliersList = require('./routes/suppliers/suppliers-list'); and app.use('/api', suppliersList); then I created my API and my tests. I will commit. Oh there was an error that occurred in the CORS, it was res.setHeader('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, PATCH, OPTIONS'); PATCH was highlighted in RED so i deleted it and it got rid of the error.